+++ src/core/state/components/StoreProvider.tsx
@@ -1,10 +1,11 @@
 import { type Store } from '@reduxjs/toolkit';
 import { type FC, useEffect, useState } from 'react';
-import { ActivityIndicator } from 'react-native';
 import { type Persistor } from 'redux-persist';
 
 import { getMMKVEncryptionKey } from '../keychain.ts';
 import { generateStore } from '../store.ts';
+
+import AppLoading from '@/components/AppLoading.tsx';
 
 interface StoreDetails {
   store: Store;
@@ -33,7 +34,7 @@ export const StoreProvider: FC<IProps> = ({ children }) => {
   }, []);
 
   if (!storeDetails) {
-    return <ActivityIndicator size="large" color="#000000" />;
+    return <AppLoading />;
   }
 
   return children(storeDetails);
